densapp/requirements.txt
asgiref==3.8.1
Authlib==1.0.1
certifi==2024.12.14
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cryptography==3.4.8
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-jwt==1.13.4
gunicorn==23.0.0
h11==0.14.0
idna==2.8
nh3==0.2.19
oauthlib==3.2.2
packaging==24.2
psycopg==3.2.1
psycopg-binary==3.2.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tinycss2==1.1.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
